abi: Use uintptr_t for __syscall() arguments.

+++ libraries/libsystem/plugs/__plug_filesystem.cpp
@@ -5,25 +5,25 @@
 
 Result filesystem_link(const char *oldpath, const char *newpath)
 {
-    return __syscall(SYS_FILESYSTEM_LINK, (int)oldpath, (int)newpath);
+    return __syscall(SYS_FILESYSTEM_LINK, (uintptr_t)oldpath, (uintptr_t)newpath);
 }
 
 Result filesystem_unlink(const char *path)
 {
-    return __syscall(SYS_FILESYSTEM_UNLINK, (int)path);
+    return __syscall(SYS_FILESYSTEM_UNLINK, (uintptr_t)path);
 }
 
 Result filesystem_mkdir(const char *path)
 {
-    return __syscall(SYS_FILESYSTEM_MKDIR, (int)path);
+    return __syscall(SYS_FILESYSTEM_MKDIR, (uintptr_t)path);
 }
 
 Result filesystem_mkpipe(const char *path)
 {
-    return __syscall(SYS_FILESYSTEM_MKPIPE, (int)path);
+    return __syscall(SYS_FILESYSTEM_MKPIPE, (uintptr_t)path);
 }
 
 Result filesystem_rename(const char *old_path, const char *new_path)
 {
-    return __syscall(SYS_FILESYSTEM_RENAME, (int)old_path, (int)new_path);
+    return __syscall(SYS_FILESYSTEM_RENAME, (uintptr_t)old_path, (uintptr_t)new_path);
 }
