+++ libraries/libsystem/plugs/__plug_memory.cpp
@@ -7,7 +7,7 @@
 
 Result memory_alloc(size_t size, uintptr_t *out_address)
 {
-    return __syscall(SYS_MEMORY_ALLOC, size, (int)out_address);
+    return __syscall(SYS_MEMORY_ALLOC, size, (uintptr_t)out_address);
 }
 
 Result memory_free(uintptr_t address)
@@ -17,10 +17,10 @@ Result memory_free(uintptr_t address)
 
 Result memory_include(int handle, uintptr_t *out_address, size_t *out_size)
 {
-    return __syscall(SYS_MEMORY_INCLUDE, handle, (int)out_address, (int)out_size);
+    return __syscall(SYS_MEMORY_INCLUDE, handle, (uintptr_t)out_address, (uintptr_t)out_size);
 }
 
 Result memory_get_handle(uintptr_t address, int *out_handle)
 {
-    return __syscall(SYS_MEMORY_GET_HANDLE, address, (int)out_handle);
+    return __syscall(SYS_MEMORY_GET_HANDLE, address, (uintptr_t)out_handle);
 }
