+++ libraries/libsystem/plugs/__plug_process.cpp
@@ -10,7 +10,7 @@ int __plug_process_this()
 {
     if (_cached_this == -1)
     {
-        assert(__syscall(SYS_PROCESS_THIS, (int)&_cached_this) == SUCCESS);
+        assert(__syscall(SYS_PROCESS_THIS, (uintptr_t)&_cached_this) == SUCCESS);
     }
 
     return _cached_this;
@@ -18,7 +18,7 @@ int __plug_process_this()
 
 Result __plug_process_launch(Launchpad *launchpad, int *pid)
 {
-    return __syscall(SYS_PROCESS_LAUNCH, (int)launchpad, (int)pid);
+    return __syscall(SYS_PROCESS_LAUNCH, (uintptr_t)launchpad, (uintptr_t)pid);
 }
 
 void __plug_process_exit(int code)
@@ -35,12 +35,12 @@ Result __plug_process_cancel(int pid)
 
 Result __plug_process_get_directory(char *buffer, uint size)
 {
-    return __syscall(SYS_PROCESS_GET_DIRECTORY, (int)buffer, size);
+    return __syscall(SYS_PROCESS_GET_DIRECTORY, (uintptr_t)buffer, size);
 }
 
 Result __plug_process_set_directory(const char *directory)
 {
-    return __syscall(SYS_PROCESS_SET_DIRECTORY, (int)directory);
+    return __syscall(SYS_PROCESS_SET_DIRECTORY, (uintptr_t)directory);
 }
 
 Result __plug_process_sleep(int time)
@@ -50,5 +50,5 @@ Result __plug_process_sleep(int time)
 
 Result __plug_process_wait(int pid, int *exit_value)
 {
-    return __syscall(SYS_PROCESS_WAIT, pid, (int)exit_value);
+    return __syscall(SYS_PROCESS_WAIT, pid, (uintptr_t)exit_value);
 }
