+++ libraries/libsystem/plugs/__plug_system.cpp
@@ -5,24 +5,24 @@
 
 void __plug_system_get_info(SystemInfo *info)
 {
-    __syscall(SYS_SYSTEM_GET_INFO, (int)info);
+    __syscall(SYS_SYSTEM_GET_INFO, (uintptr_t)info);
 }
 
 void __plug_system_get_status(SystemStatus *status)
 {
-    __syscall(SYS_SYSTEM_GET_STATUS, (int)status);
+    __syscall(SYS_SYSTEM_GET_STATUS, (uintptr_t)status);
 }
 
 TimeStamp __plug_system_get_time()
 {
     TimeStamp timestamp = 0;
-    __syscall(SYS_SYSTEM_GET_TIME, (int)&timestamp);
+    __syscall(SYS_SYSTEM_GET_TIME, (uintptr_t)&timestamp);
     return timestamp;
 }
 
 uint __plug_system_get_ticks()
 {
     uint result = 0;
-    __syscall(SYS_SYSTEM_GET_TICKS, (int)&result);
+    __syscall(SYS_SYSTEM_GET_TICKS, (uintptr_t)&result);
     return result;
 }
